Stringify byte buffers in diffy avro keys
The default toString for a byte buffer doesn't contain its content,
 just metadata that's the same for every byte buffer of the same size.
If you used a bytes field in Avro as your key,
this would cause a hot key of that default string, rather than grouping by key values.

This fix uses the UTF-8 version of what's actually inside the buffer.