Add logical decimal type support to avroOf
Attempt to add support for logical decimal types to avroOf
See issue #178

It appears the Builder still fails when nested records rely on Avro
Conversions since it only includes them in the generated classes for the
nested record, and not for the parent record builder.

I also could not get it to work for nullable logical decimal types as
the type coercion fails for the null array (it's a bit related to the
above issue that we'd want it to call the Conversion and also handle
nulls - e.g. 0 byte array -> null).

Note I also don't handle the case of a fixed underlying byte array size
in the Avro schema, nor checking that the generated integer will fit
inside the decimal's precision and scale - but at the moment the above
issues are more blocking.